First pass at moving put requests to form
